Move `TemporalUtilities` to the metadata modules since it is not needed by the utility modules. This change has an impact on the services that the module declares to use.
