+++ core/sis-metadata/src/main/java/org/apache/sis/internal/metadata/TemporalUtilities.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.sis.internal.util;
+package org.apache.sis.internal.metadata;
 
 import java.util.Date;
 import java.util.ServiceLoader;
