+++ storage/sis-geotiff/src/main/java/org/apache/sis/storage/geotiff/XMLMetadata.java
@@ -47,7 +47,7 @@ import org.apache.sis.util.collection.TableColumn;
 import org.apache.sis.util.resources.Errors;
 import org.apache.sis.xml.XML;
 
-import static org.apache.sis.internal.util.TemporalUtilities.toDate;
+import static org.apache.sis.internal.metadata.TemporalUtilities.toDate;
 
 
 /**
